+++ src/main/scala/ai/tripl/arc/ARC.scala
@@ -223,7 +223,7 @@ object ARC {
       activeLifecyclePlugins=Nil,
       pipelineStagePlugins=ServiceLoader.load(classOf[PipelineStagePlugin], loader).iterator().asScala.toList,
       udfPlugins=ServiceLoader.load(classOf[UDFPlugin], loader).iterator().asScala.toList,
-      serializableConfiguration=Option(new SerializableConfiguration(spark.sparkContext.hadoopConfiguration)),
+      serializableConfiguration=new SerializableConfiguration(spark.sparkContext.hadoopConfiguration),
       userData=collection.mutable.Map.empty
     )
 
@@ -237,7 +237,7 @@ object ARC {
         .field("event", "enter")
         .field("config", sparkConf)
         .field("sparkVersion", spark.version)
-        .field("frameworkVersion", frameworkVersion)
+        .field("arcVersion", frameworkVersion)
         .field("hadoopVersion", org.apache.hadoop.util.VersionInfo.getVersion)
         .field("scalaVersion", scala.util.Properties.versionNumberString)
         .field("javaVersion", System.getProperty("java.runtime.version"))
@@ -267,7 +267,7 @@ object ARC {
         logger.error()
           .field("event", "exit")
           .field("sparkVersion", spark.version)
-          .field("frameworkVersion", frameworkVersion)
+          .field("arcVersion", frameworkVersion)
           .field("hadoopVersion", org.apache.hadoop.util.VersionInfo.getVersion)
           .field("scalaVersion", scala.util.Properties.versionNumberString)
           .field("javaVersion", System.getProperty("java.runtime.version"))