+++ src/main/scala/ai/tripl/arc/load/TextLoad.scala
@@ -155,7 +155,7 @@ object TextLoadStage {
 
         repartitionedDF.foreachPartition { partition: Iterator[Row] =>
           if (partition.hasNext) {
-            val hadoopConf = arcContext.serializableConfiguration.get.value
+            val hadoopConf = arcContext.serializableConfiguration.value
 
             // buffer so first row can be accessed
             val bufferedPartition = partition.buffered