+++ src/main/scala/ai/tripl/arc/util/CloudUtils.scala
@@ -19,7 +19,7 @@ object CloudUtils {
   // org.apache.hadoop.fs.s3a.AnonymousAWSCredentialsProvider to support anonymous credentials
   val defaultAWSProvidersOverride = "org.apache.hadoop.fs.s3a.SimpleAWSCredentialsProvider,com.amazonaws.auth.EnvironmentVariableCredentialsProvider,com.amazonaws.auth.InstanceProfileCredentialsProvider,com.amazonaws.auth.ContainerCredentialsProvider,org.apache.hadoop.fs.s3a.AnonymousAWSCredentialsProvider"
 
-  def setHadoopConfiguration(authentication: Option[API.Authentication])(implicit spark: SparkSession, logger: ai.tripl.arc.util.log.logger.Logger, arcContext: ARCContext) = {
+  def setHadoopConfiguration(authentication: Option[API.Authentication])(implicit spark: SparkSession, logger: ai.tripl.arc.util.log.logger.Logger) = {
     import spark.sparkContext.{hadoopConfiguration => hc}
 
     // clear s3a settings
@@ -148,9 +148,6 @@ object CloudUtils {
       }
       case None =>
     }
-
-    // update the serializableConfiguration
-    arcContext.serializableConfiguration = Option(new SerializableConfiguration(spark.sparkContext.hadoopConfiguration))
   }
 
   // using a string filePath so that invalid paths can be used for local files