+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.10.1
 
 - add the `get_uri` user defined function which returns the contents of a URI as an `Array[Byte]` which can be used with `decode` to convert to text.
+- rename `frameworkVersion` to `arcVersion` in initiliasation logs for clarity.
 
 ## 2.10.0